Velum: Drop --timeout option for Velum bootstrap
The step is wrapped with the timeout option so passing timeout is not
needed. We use 10 minutes instead of 125 which is the default for Velum.
Previously it was broken because the timeout() option was set to 125
minutes but we were passing --timeout 2 to the script so we ended up
waiting only 2 minutes which may not be enough for certain environments.